Move functions to LoaderUtils
