--- src/core/renderer/utilities/arrayToString.js
@@ -1,7 +1,0 @@
-const utf8decoder = new TextDecoder();
-
-export function arrayToString( array ) {
-
-	return utf8decoder.decode( array );
-
-}
